# (feature) Add memory pooling to CLOB module, latency reductions of 30-40% in benchmarking