--- a/protocol/x/clob/memclob/memclob.go
+++ b/protocol/x/clob/memclob/memclob.go
@@ -1736,8 +1736,8 @@ func (m *MemClobPriceTimePriority) mustPerformTakerOrderMatching(
 
 		// Pass in empty map to avoid reading `AffiliateWhitelist` from state in every `CheckTx`. This deviates
 		// from `DeliverTx` which accounts for affiliate whitelist correctly. This deviation is ok because rev
-		// shares/fees are distributed to the recipient’s bank balance and not settled at the subaccount level,
-		// and won’t affect the collateralization of future operations in the operations queue.
+		// shares/fees are distributed to the recipient's bank balance and not settled at the subaccount level,
+		// and won't affect the collateralization of future operations in the operations queue.
 		success, takerUpdateResult, makerUpdateResult, _, err := m.clobKeeper.ProcessSingleMatch(
 			ctx, &matchWithOrders, map[string]uint32{})
 		if err != nil && !errors.Is(err, satypes.ErrFailedToUpdateSubaccounts) {
@@ -1852,10 +1852,11 @@ func (m *MemClobPriceTimePriority) mustPerformTakerOrderMatching(
 		}
 
 		// 3.
-		newMakerFills = append(newMakerFills, types.MakerFill{
-			MakerOrderId: makerOrderId,
-			FillAmount:   matchedAmount.ToUint64(),
-		})
+		// Use the memory pool to get a MakerFill object instead of creating a new one
+		makerFill := GlobalMemPools.MakerFillPool.Get()
+		makerFill.MakerOrderId = makerOrderId
+		makerFill.FillAmount = matchedAmount.ToUint64()
+		newMakerFills = append(newMakerFills, *makerFill)
 
 		// 4.
 		if newTakerOrder.IsReduceOnly() && takerRemainingSize > 0 {